backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
redis
Pillow
django-cors-headers